backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.37
psycopg==3.2.4
psycopg-binary==3.2.4
pydantic==2.10.5
python-dotenv==1.0.1
starlette==0.41.3
httpx==0.28.1
anyio==4.8.0